Moved files around and further implemented cli
