+++ .gitignore
@@ -24,3 +24,4 @@ Cargo.lock
 
 # Project specific files
 temp.json
+real.json
